Moving Telemetry stuff into its own namespace
