+++ src/Microsoft.Identity.Client/Internal/Requests/RequestBase.cs
@@ -33,6 +33,7 @@ using System.Threading.Tasks;
 using Microsoft.Identity.Client.Internal.Cache;
 using Microsoft.Identity.Client.Internal.Instance;
 using Microsoft.Identity.Client.Internal.OAuth2;
+using Microsoft.Identity.Client.Internal.Telemetry;
 
 namespace Microsoft.Identity.Client.Internal.Requests
 {
@@ -87,7 +88,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
             SupportADFS = false;
 
             AuthenticationRequestParameters.LogState();
-            Telemetry.GetInstance().ClientId = AuthenticationRequestParameters.ClientId;
+            Client.Telemetry.GetInstance().ClientId = AuthenticationRequestParameters.ClientId;
         }
 
         protected virtual SortedSet<string> GetDecoratedScope(SortedSet<string> inputScope)
@@ -116,7 +117,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
         public async Task<AuthenticationResult> RunAsync()
         {
             //this method is the common entrance for all token requests, so it is a good place to put the generic Telemetry logic here
-            AuthenticationRequestParameters.RequestContext.TelemetryRequestId = Telemetry.GetInstance().GenerateNewRequestId();
+            AuthenticationRequestParameters.RequestContext.TelemetryRequestId = Client.Telemetry.GetInstance().GenerateNewRequestId();
             var apiEvent = new ApiEvent()
             {
                 ApiId = ApiId,
@@ -135,7 +136,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
                 apiEvent.AuthorityType = AuthenticationRequestParameters.Authority.AuthorityType.ToString();
             }
 
-            Telemetry.GetInstance().StartEvent(AuthenticationRequestParameters.RequestContext.TelemetryRequestId, apiEvent);
+            Client.Telemetry.GetInstance().StartEvent(AuthenticationRequestParameters.RequestContext.TelemetryRequestId, apiEvent);
 
             try
             {
@@ -157,8 +158,8 @@ namespace Microsoft.Identity.Client.Internal.Requests
             }
             finally
             {
-                Telemetry.GetInstance().StopEvent(AuthenticationRequestParameters.RequestContext.TelemetryRequestId, apiEvent);
-                Telemetry.GetInstance().Flush(AuthenticationRequestParameters.RequestContext.TelemetryRequestId);
+                Client.Telemetry.GetInstance().StopEvent(AuthenticationRequestParameters.RequestContext.TelemetryRequestId, apiEvent);
+                Client.Telemetry.GetInstance().Flush(AuthenticationRequestParameters.RequestContext.TelemetryRequestId);
             }
         }
