+++ src/Microsoft.Identity.Client/Internal/Telemetry/UiEvent.cs
@@ -26,7 +26,7 @@
 //------------------------------------------------------------------------------
 
 
-namespace Microsoft.Identity.Client
+namespace Microsoft.Identity.Client.Internal.Telemetry
 {
     internal class UiEvent : EventBase
     {
