+++ tests/Test.MSAL.NET.Unit/PublicClientApplicationTests.cs
@@ -40,6 +40,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NSubstitute;
 using Test.MSAL.NET.Unit.Mocks;
 using Microsoft.Identity.Client.Internal.OAuth2;
+using Microsoft.Identity.Client.Internal.Telemetry;
 
 namespace Test.MSAL.NET.Unit
 {
