+++ tests/Test.MSAL.NET.Unit/RequestsTests/InteractiveRequestTests.cs
@@ -41,6 +41,7 @@ using Microsoft.Identity.Client.Internal.Requests;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NSubstitute;
 using Test.MSAL.NET.Unit.Mocks;
+using Microsoft.Identity.Client.Internal.Telemetry;
 
 namespace Test.MSAL.NET.Unit.RequestsTests
 {
