+++ tests/Test.MSAL.NET.Unit/TelemetryTests.cs
@@ -29,6 +29,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Identity.Client;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.Identity.Client.Internal.Telemetry;
 
 namespace Test.MSAL.NET.Unit
 {
